requirements.txt
uvicorn
aiohttp
aioredis[hiredis]
blacksheep
jinja2
orjson
python3-openid
python-slugify
